go.mod: Clarify retractions
Add a comment to go.mod that is placed according to the go.mod
specification to give some background to the retraction of older
releases. Also make the text more clear to explain what happened.

The 2.0.0 version is removed from the list, as it was leading to an
error e.g. from go mod tidy:

    go.mod:24:2: retract github.com/tredoe/osutil: version "v2.0.0" invalid: should be v0 or v1, not v2